Prepare for template inheritance static analysis.
